requirements.txt
flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==20.1.0
google-generativeai==0.3.1
requests==2.31.0
uuid==1.30
